+++ scripts/allegrovideo.py
@@ -2,7 +2,7 @@ import time
 import gradio as gr
 import transformers
 import diffusers
-from modules import scripts, processing, shared, images, devices, sd_models, sd_checkpoint, model_quant, timer
+from modules import scripts, processing, shared, images, devices, sd_models, sd_checkpoint, model_quant, timer, sd_hijack_te
 
 
 repo_id = 'rhymes-ai/Allegro'
@@ -16,16 +16,6 @@ def hijack_decode(*args, **kwargs):
     t1 = time.time()
     timer.process.add('vae', t1-t0)
     shared.log.debug(f'Video: vae={vae.__class__.__name__} time={t1-t0:.2f}')
-    return res
-
-
-def hijack_encode_prompt(*args, **kwargs):
-    t0 = time.time()
-    res = shared.sd_model.orig_encode_prompt(*args, **kwargs)
-    t1 = time.time()
-    timer.process.add('te', t1-t0)
-    shared.log.debug(f'Video: te={shared.sd_model.text_encoder.__class__.__name__} time={t1-t0:.2f}')
-    shared.sd_model = sd_models.apply_balanced_offload(shared.sd_model)
     return res
 
 
@@ -94,9 +84,9 @@ class Script(scripts.Script):
             shared.sd_model.vae.orig_decode = shared.sd_model.vae.decode
             shared.sd_model.orig_encode_prompt = shared.sd_model.encode_prompt
             shared.sd_model.vae.decode = hijack_decode
-            shared.sd_model.encode_prompt = hijack_encode_prompt
             shared.sd_model.vae.enable_tiling()
             # shared.sd_model.vae.enable_slicing()
+            sd_hijack_te.init_hijack(shared.sd_model)
 
         shared.sd_model = sd_models.apply_balanced_offload(shared.sd_model)
         devices.torch_gc(force=True)